backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
PyMySQL>=1.0
python-dotenv>=1.0
djangorestframework-simplejwt>=5.2
